pyproject.toml
[tool.poetry]
name = "aihbridge"
version = "0.1.0"
description = ""
authors = ["AyaMoeen <ayamoinn95@gmail.com>"]
package-mode = false 
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
ruff = "^0.12.9"
pytest = "^8.4.1"
djangorestframework = "^3.16.1"
python-decouple = "^3.8"
django-cors-headers = "^4.7.0"
pillow = "^11.3.0"
requests = "^2.32.5"


[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
djangorestframework-stubs = "^3.16.2"
types-requests = "^2.32.4.20250809"
django-stubs = "^5.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
